performed extract class on configconstants.java
