+++ server/src/main/java/cn/keking/config/configconstants/FtpConfigConstants.java
@@ -1,4 +1,4 @@
-package cn.keking.config;
+package cn.keking.config.configconstants;
 
 import org.springframework.beans.factory.annotation.Value;
 
@@ -7,10 +7,24 @@ public class FtpConfigConstants
     private static String ftpUsername;
     private static String ftpPassword;
     private static String ftpControlEncoding;
+    private static Boolean fileUploadDisable;
 
     public static final String DEFAULT_FTP_USERNAME = null;
     public static final String DEFAULT_FTP_PASSWORD = null;
     public static final String DEFAULT_FTP_CONTROL_ENCODING = "UTF-8";
+
+    public static void setFileUploadDisableValue(Boolean fileUploadDisable) {
+        FtpConfigConstants.fileUploadDisable = fileUploadDisable;
+    }
+
+    public static Boolean getFileUploadDisable() {
+        return fileUploadDisable;
+    }
+
+    @Value("${file.upload.disable:false}")
+    public static void setFileUploadDisable(Boolean fileUploadDisable) {
+        setFileUploadDisableValue(fileUploadDisable);
+    }
 
     public static String getFtpUsername() {
         return ftpUsername;
