+++ server/src/main/java/cn/keking/model/FileType.java
@@ -1,6 +1,6 @@
 package cn.keking.model;
 
-import cn.keking.config.ConfigConstants;
+import cn.keking.config.configconstants.ConfigConstants;
 import cn.keking.config.MediaConfigConstants;
 
 import java.util.HashMap;
