+++ server/src/main/java/cn/keking/service/impl/OfficeFilePreviewImpl.java
@@ -1,6 +1,6 @@
 package cn.keking.service.impl;
 
-import cn.keking.config.ConfigConstants;
+import cn.keking.config.configconstants.ConfigConstants;
 import cn.keking.model.FileAttribute;
 import cn.keking.model.ReturnResponse;
 import cn.keking.service.FilePreview;
