+++ server/src/main/java/cn/keking/utils/DownloadUtils.java
@@ -1,6 +1,6 @@
 package cn.keking.utils;
 
-import cn.keking.config.ConfigConstants;
+import cn.keking.config.configconstants.ConfigConstants;
 import cn.keking.model.FileAttribute;
 import cn.keking.model.ReturnResponse;
 import io.mola.galimatias.GalimatiasParseException;
